+++ src/core/ioHandler/wspsaWeightIncidence.py
