+++ src/helpers/downloadNetwork.py
